+++ packages/react/src/receipt/receipt.tsx
@@ -82,7 +82,7 @@ export function ReceiptDisplay({
     );
   }
 
-  // 2.0.0-rc1 and later use the latest component with Person type support
+  // 2.0.0-rc2 and later use the latest component with Person type support
   return (
     <ReceiptLatest
       receipt={data}