+++ packages/schema/src/lts/index.ts
@@ -3,4 +3,4 @@ export * as v1_8_0 from "./1.8.0";
 export * as v1_9_0 from "./1.9.0";
 export * as v1_10_0 from "./1.10.0";
 export * as v1_11_0 from "./1.11.0";
-export * as v2_0_0_rc1 from "./2.0.0-rc1";
+export * as v2_0_0_rc2 from "./2.0.0-rc2";