+++ packages/pdfgen/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @versaprotocol/pdfgen
 
+## 1.12.4
+
+### Patch Changes
+
+- Patch updates for schema version 2.0.0-rc2
+- Updated dependencies
+  - @versaprotocol/schema@1.12.1
+  - @versaprotocol/belt@1.12.3
+
 ## 1.12.3
 
 ### Patch Changes